refactor(test): split implicit-resume-from-derived-key.test.ts (#628) (#728)
